requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
boto3==1.40.16
botocore==1.40.16
certifi==2025.8.3
deprecation==2.1.0
dj-database-url==3.0.1
Django==5.2.5
django-ckeditor==6.7.3
django-js-asset==3.1.2
django-storages==1.14.6
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jmespath==1.0.1
packaging==25.0
pillow==11.3.0
postgrest==1.1.1
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
realtime==2.7.0
s3transfer==0.13.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
storage3==0.12.1
StrEnum==0.4.15
supabase==2.18.1
supabase_auth==2.12.3
supabase_functions==0.10.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
websockets==15.0.1
whitenoise==6.9.0
